Separate utils/linux.go into dedicated package.
