+++ linuxUtils/linux.go
@@ -1,5 +1,5 @@
 //
-// Copyright 2020 - 2022 Nestybox, Inc.
+// Copyright 2020 - 2023 Nestybox, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-package utils
+package linuxUtils
 
 import (
 	"bufio"
